cdk/requirements.txt
-e .
aws-cdk.aws-ec2==1.97.0
